♻️refactor: 특성 기능을 멤버 도메인에서 분리
