+++ src/main/java/com/project/likelion13th_team1/domain/feature/converter/FeatureConverter.java
@@ -1,0 +1,27 @@
+package com.project.likelion13th_team1.domain.feature.converter;
+
+import com.project.likelion13th_team1.domain.feature.dto.request.FeatureRequestDto;
+import com.project.likelion13th_team1.domain.feature.dto.response.FeatureResponseDto;
+import com.project.likelion13th_team1.domain.feature.entity.Feature;
+
+public class FeatureConverter {
+
+    public static Feature toFeature(FeatureRequestDto.FeatureCreateRequestDto dto) {
+        return Feature.builder()
+                .q1(dto.q1())
+                .q2(dto.q2())
+                .q3(dto.q3())
+                .q4(dto.q4())
+                .build();
+    }
+
+    public static FeatureResponseDto.FeatureDetailResponseDto toFeatureDetailResponseDto(Feature feature) {
+        return FeatureResponseDto.FeatureDetailResponseDto.builder()
+                .featureId(feature.getId())
+                .q1(feature.getQ1())
+                .q2(feature.getQ2())
+                .q3(feature.getQ3())
+                .q4(feature.getQ4())
+                .build();
+    }
+}
