+++ src/main/java/com/project/likelion13th_team1/domain/feature/dto/request/FeatureRequestDto.java
@@ -1,0 +1,20 @@
+package com.project.likelion13th_team1.domain.feature.dto.request;
+
+public class FeatureRequestDto {
+
+    public record FeatureCreateRequestDto(
+            Integer q1,
+            Integer q2,
+            Integer q3,
+            Integer q4
+    ) {
+    }
+
+    public record FeatureUpdateRequestDto(
+            Integer q1,
+            Integer q2,
+            Integer q3,
+            Integer q4
+    ) {
+    }
+}
