+++ src/main/java/com/project/likelion13th_team1/domain/feature/dto/response/FeatureResponseDto.java
@@ -1,0 +1,16 @@
+package com.project.likelion13th_team1.domain.feature.dto.response;
+
+import lombok.Builder;
+
+public class FeatureResponseDto {
+
+    @Builder
+    public record FeatureDetailResponseDto(
+            Long featureId,
+            Integer q1,
+            Integer q2,
+            Integer q3,
+            Integer q4
+    ) {
+    }
+}
