+++ src/main/java/com/project/likelion13th_team1/domain/feature/entity/Feature.java
@@ -1,7 +1,6 @@
-package com.project.likelion13th_team1.global.feature.entity;
+package com.project.likelion13th_team1.domain.feature.entity;
 
 
-import com.project.likelion13th_team1.global.feature.dto.request.FeatureRequestDto;
 import jakarta.persistence.*;
 import lombok.*;
 
@@ -32,10 +31,20 @@ public class Feature {
     @Column(name = "Q4")
     private Integer q4;
 
-    public void updateFeature(FeatureRequestDto.FeatureUpdateRequestDto dto) {
-        if (dto.messyHouse() != null) this.messyHouse = dto.messyHouse();
-        if (dto.cleaningStyle() != null) this.cleaningStyle = dto.cleaningStyle();
-        if (dto.cleanHouse() != null) this.cleanHouse = dto.cleanHouse();
-        if (dto.routineStyle() != null) this.routineStyle = dto.routineStyle();
+
+    public void updateQ1(Integer q1) {
+        this.q1 = q1;
+    }
+
+    public void updateQ2(Integer q2) {
+        this.q2 = q2;
+    }
+
+    public void updateQ3(Integer q3) {
+        this.q3 = q3;
+    }
+
+    public void updateQ4(Integer q4) {
+        this.q4 = q4;
     }
 }
