+++ src/main/java/com/project/likelion13th_team1/domain/feature/exception/FeatureErrorCode.java
@@ -1,4 +1,4 @@
-package com.project.likelion13th_team1.global.feature.exception;
+package com.project.likelion13th_team1.domain.feature.exception;
 
 import com.project.likelion13th_team1.global.apiPayload.code.BaseErrorCode;
 import lombok.AllArgsConstructor;
