+++ src/main/java/com/project/likelion13th_team1/domain/feature/exception/FeatureException.java
@@ -1,6 +1,5 @@
-package com.project.likelion13th_team1.global.feature.exception;
+package com.project.likelion13th_team1.domain.feature.exception;
 
-import com.project.likelion13th_team1.domain.member.exception.MemberErrorCode;
 import com.project.likelion13th_team1.global.apiPayload.exception.CustomException;
 
 public class FeatureException extends CustomException {
