+++ src/main/java/com/project/likelion13th_team1/domain/feature/repository/FeatureRepository.java
@@ -1,6 +1,6 @@
-package com.project.likelion13th_team1.global.feature.repository;
+package com.project.likelion13th_team1.domain.feature.repository;
 
-import com.project.likelion13th_team1.global.feature.entity.Feature;
+import com.project.likelion13th_team1.domain.feature.entity.Feature;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FeatureRepository extends JpaRepository<Feature, Long> {
