+++ src/main/java/com/project/likelion13th_team1/domain/feature/service/query/FeatureQueryServiceImpl.java
@@ -1,0 +1,34 @@
+package com.project.likelion13th_team1.domain.feature.service.query;
+
+import com.project.likelion13th_team1.domain.member.entity.Member;
+import com.project.likelion13th_team1.domain.member.exception.MemberErrorCode;
+import com.project.likelion13th_team1.domain.member.exception.MemberException;
+import com.project.likelion13th_team1.domain.member.repository.MemberRepository;
+import com.project.likelion13th_team1.domain.feature.converter.FeatureConverter;
+import com.project.likelion13th_team1.domain.feature.dto.response.FeatureResponseDto;
+import com.project.likelion13th_team1.domain.feature.entity.Feature;
+import com.project.likelion13th_team1.domain.feature.exception.FeatureErrorCode;
+import com.project.likelion13th_team1.domain.feature.exception.FeatureException;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@Transactional
+@RequiredArgsConstructor
+public class FeatureQueryServiceImpl implements FeatureQueryService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public FeatureResponseDto.FeatureDetailResponseDto getFeature(String email) {
+        Member member = memberRepository.findByEmail(email)
+                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
+
+        Feature feature = member.getFeature();
+        if(feature == null) {
+            throw new FeatureException(FeatureErrorCode.FEATURE_NOT_FOUND);
+        }
+        return FeatureConverter.toFeatureDetailResponseDto(feature);
+    }
+}
