+++ src/main/java/com/project/likelion13th_team1/domain/member/controller/MemberController.java
@@ -5,8 +5,6 @@ import com.project.likelion13th_team1.domain.member.dto.response.MemberResponseD
 import com.project.likelion13th_team1.domain.member.service.command.MemberCommandService;
 import com.project.likelion13th_team1.domain.member.service.query.MemberQueryService;
 import com.project.likelion13th_team1.global.apiPayload.CustomResponse;
-import com.project.likelion13th_team1.global.feature.dto.request.FeatureRequestDto;
-import com.project.likelion13th_team1.global.feature.dto.response.FeatureResponseDto;
 import com.project.likelion13th_team1.global.security.userdetails.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -62,31 +60,4 @@ public class MemberController {
         return CustomResponse.onSuccess(HttpStatus.NO_CONTENT, "회원 탈퇴 완료");
     }
 
-    @Operation(summary = "회원 특성 정보 생성", description = "루틴 추천을 위한 회원 특성 정보를 생성한다.")
-    @PostMapping("/feature")
-    public CustomResponse<String> createFeature(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails,
-            @RequestBody FeatureRequestDto.FeatureCreateRequestDto featureCreateRequestDto
-            ) {
-        memberCommandService.createFeature(customUserDetails.getUsername(), featureCreateRequestDto);
-        return CustomResponse.onSuccess(HttpStatus.CREATED, "회원 특성 정보 생성 완료");
-    }
-
-    @Operation(summary = "회원 특성 정보 수정", description = "루틴 추천을 위한 회원 특성 정보를 수정한다.")
-    @PatchMapping("/feature")
-    public CustomResponse<String> updateFeature(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails,
-            @RequestBody FeatureRequestDto.FeatureUpdateRequestDto featureUpdateRequestDto
-    ) {
-        memberCommandService.updateFeature(customUserDetails.getUsername(), featureUpdateRequestDto);
-        return CustomResponse.onSuccess(HttpStatus.OK, "회원 특성 정보 수정 완료");
-    }
-
-    @Operation(summary = "회원 특성 정보 조회", description = "회원 특성 정보를 조회한다.")
-    @GetMapping("/feature")
-    public CustomResponse<FeatureResponseDto.FeatureDetailResponseDto> getFeature(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails
-    ) {
-        return CustomResponse.onSuccess(memberCommandService.getFeature(customUserDetails.getUsername()));
-    }
 }
