+++ src/main/java/com/project/likelion13th_team1/domain/member/service/command/MemberCommandServiceImpl.java
@@ -6,13 +6,7 @@ import com.project.likelion13th_team1.domain.member.entity.Member;
 import com.project.likelion13th_team1.domain.member.exception.MemberErrorCode;
 import com.project.likelion13th_team1.domain.member.exception.MemberException;
 import com.project.likelion13th_team1.domain.member.repository.MemberRepository;
-import com.project.likelion13th_team1.global.feature.converter.FeatureConverter;
-import com.project.likelion13th_team1.global.feature.dto.request.FeatureRequestDto;
-import com.project.likelion13th_team1.global.feature.dto.response.FeatureResponseDto;
-import com.project.likelion13th_team1.global.feature.entity.Feature;
-import com.project.likelion13th_team1.global.feature.exception.FeatureErrorCode;
-import com.project.likelion13th_team1.global.feature.exception.FeatureException;
-import com.project.likelion13th_team1.global.feature.repository.FeatureRepository;
+import com.project.likelion13th_team1.domain.feature.repository.FeatureRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -58,41 +52,5 @@ public class MemberCommandServiceImpl implements MemberCommandService {
 
         // soft delete
         member.delete();
-    }
-
-    @Override
-    public void createFeature(String email, FeatureRequestDto.FeatureCreateRequestDto featureCreateRequestDto) {
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-
-        // TODO : 중복저장 방지
-        Feature feature = FeatureConverter.toFeature(featureCreateRequestDto, FeatureType.MEMBER);
-        featureRepository.save(feature);
-
-        member.linkFeature(feature);
-    }
-
-    @Override
-    public void updateFeature(String email, FeatureRequestDto.FeatureUpdateRequestDto featureUpdateRequestDto) {
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-
-        Feature feature = member.getFeature();
-        if(feature == null) {
-            throw new FeatureException(FeatureErrorCode.FEATURE_NOT_FOUND);
-        }
-        feature.updateFeature(featureUpdateRequestDto);
-    }
-
-    @Override
-    public FeatureResponseDto.FeatureDetailResponseDto getFeature(String email) {
-        Member member = memberRepository.findByEmail(email)
-                .orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-
-        Feature feature = member.getFeature();
-        if(feature == null) {
-            throw new FeatureException(FeatureErrorCode.FEATURE_NOT_FOUND);
-        }
-        return FeatureConverter.toFeatureDetailResponseDto(feature);
     }
 }
