+++ src/main/java/com/project/likelion13th_team1/domain/routine/entity/Routine.java
@@ -2,16 +2,12 @@ package com.project.likelion13th_team1.domain.routine.entity;
 
 import com.project.likelion13th_team1.domain.event.entity.Event;
 import com.project.likelion13th_team1.domain.group.entity.Group;
-import com.project.likelion13th_team1.domain.member.entity.Member;
 import com.project.likelion13th_team1.domain.routine.dto.request.RoutineRequestDto;
 import com.project.likelion13th_team1.global.entity.BaseEntity;
-import com.project.likelion13th_team1.global.entity.Status;
-import com.project.likelion13th_team1.global.feature.entity.Feature;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
